Cleanup & split tests into unit and integration groups
