--- babel.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-    // presets: [['@babel/preset-env', { targets: { node: 'current' } }], '@babel/preset-typescript'],
-    // plugins: ['replace-ts-export-assignment'],
-};
