--- src/__test__/mocha.opts
@@ -1,5 +1,0 @@
---require ts-node/register
---require @babel/register
-
-
-src/__test__/**/*.spec.ts
